REMOTE_LIBRARY1 - re-implemented remoteArtisanLibrary
This change had several parts.  Some were just cosmetic/aestheic

- changed $this_uuid to 'ArtisanPerl-MACHINE_NAME', which was not technically needed, for aesthetic purposes.
- changed name of remoteArtisan.pm to remoteArtisanLibrary.pm

The gist of the change was to allow the webUI to get information
about devices (libraries) given just their uuid, and to use that
information.

- added ip:port to localRenderer.pm for consistency and
  potential future artisanRemoteRencerer.pm
- added ip, port, local, and artisan_remote to the record
  returned to JS from webUI::getDevicesData()
- added that 'rec' to JS Device Buttons so they persistently
  contain the ip, port, local, and artisan_remote fields

Then the main change was to implement JS current_library_url()
and library_url(library_uuid) so that calls to artistanRemoteLibraries
can be detected and made directly to the given http://server_ip:server_port.
This effectively means that only Libraries that are 'online' can
be accessed, and required some error checking if if library_url()
returns ''.

This obviated the need for any explicit handling of remote_artisan
in Perl webUI.pm or the remoteArtisanLibrary.pm itself. The old code
is now commented out.

Both the Perl and JS continue to create the art_uri for Tracks
from the track's parent_id as a call to the correct server at
/get_art/parent_id/folder.jpg

A cosmetic additional change was to renderer.js to show the
library that a given track originated from, since we can now
do that.